requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-environ==0.10.0
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==1.7.1
PyMySQL==1.0.3
pytz==2023.2
sqlparse==0.2.4
tzdata==2023.2
